OFStateManager: fix bug when marking for deletion
ft_entry_mark_deleted always set the state to DELETE_MARKED whether or not the
flow was stable, so we'd always make the call into forwarding immediately (even
if the previous state was MODIFYING, meaning we'd have two requests in flight
for the same flow!).

The fix is to only send the call to forwarding if the flow was stable before
marking it for deletion.