OFStateManager: fix barrier tracking for flow deletes
Previously, the delete message handlers would free the LOCI message immediately
without waiting for Forwarding to signal that the deletion was complete. This
meant pending deletions would not block a barrier.

This change passes the usual ptr_cxn_wrapper_t as the cookie to
indigo_fwd_flow_delete instead of just the entry pointer. The ptr_cxn retains
the LOCI request for barrier tracking.

The queued request code is changed significantly. Previously (contrary to the
comments) a deletion was not tracked on the queued_reqs list. Now the ptr_cxn
is enqueued and any existing queued modifies are deleted.