OFStateManager: flowtable registration infrastructure
See the comments in indigo/of_state_manager.h. This change pulls some of the
table dispatch logic out of Forwarding and into OFStateManager. It does not yet
break the old Forwarding interface.

The plan is to do this for most "objects" in the switch (including groups,
ports, and queues).  This gives Indigo the ability to enumerate them,
simplifying the platform-dependent code and removing the need for clunky APIs
like `indigo_port_stats_get` that assume the reply fits in a single multipart
message.