retain ownership of incoming messages in OFConnectionManager
This simplifies the OFStateManager handlers which no longer need to free the
message when they're finished with it. Handlers that pass the message to a long
running task dup it first.

The end goal is to use the message directly from the OFConnectionManager wire
buffer without needing to copy it first.