OFConnectionManager: speed up logging
The log messages in cxn_instance.c automatically get the connection
IP/port/aux-id prepended. Even when the log isn't displayed due to the
loglevel, the cxn_ip_string function was still being called. The conditional
that checks whether the log level is enabled is inside aim_log_common, and the
result of cxn_ip_string is passed as a parameter to that varargs function.

Fixed by expanding the log format and args into the macros instead of hiding
them in cxn_ip_string.