OFStateManager: fix flow checksum use of table_id
The bsn_flow_checksum extension uses the table id to determine which table and
bucket checksums to update. The checksum update is done in ft_add, which uses
entry->table_id. However, ft_entry_create was not initializing the table_id
fields. Instead, for OF 1.1+, the table id was set after the call to
indigo_fwd_flow_create. This was to support the OF 1.0 feature where the
switch could choose which table to insert the flow into.

The new flowtable API does not support returning a table id, and no Indigo
switches use this feature. I'm removing it. The new code is simpler and
initializes table_id when the flow checksum code expects it to.