OFStateManager: parse match and minimatch only once when adding a flow
The match is parsed and minimatch created in flow_mod_setup_query. It was also
done again in ft_entry_create. Both of these operations are expensive, so it'd
be better to do it only once.

This change moves out the minimatch after the struct match query is done, and
moves it in to the ft_entry.