OFStateManager: throttle flow expiration
When a flow is expired it may result in sending an async message to the
controller(s). If too many flows are expired in a short period of time the
connection write queue will overflow and lead to drops. Fix this by using a
simple timer instead of the long running task, limiting the expiration rate to
around 30/s.