minimatch: cover multiple bitmap words in the utest
The utest previously only used the first 12 bytes of the of_match_t. We have to
use very few bytes in order to limit the search space of the randomized utest -
otherwise we'd never generate matches that overlap, are more-specific, etc.

This change uses several hand-picked offsets instead of the first 12 bytes.
Some of these offsets are at bitmap word boundaries, and half are in the second
bitmap word. This allows the utest to catch bugs that only occur when the
second bitmap word is used.