OFStateManager: remove support for expiring flows in Forwarding
None of our switches do this, and a switch that did would have to reimplement a
bunch of expiration.c to ratelimit the flow-removed messages and send flow-idle
notifications when necessary.

Once we're fully transitioned to the new flowtable APIs, ind_core_flow_removed
would be the only remaining user of the flow-id hashtable.