OFConnectionManager: add notifications for barriers
Some modules may wish to defer work until a barrier is received. This new
interface allows them to register to be notify when a (blocked) barrier arrives
on a connection. The module should then decide whether it holds a barrier
blocker on that connection, and if so finish its work and unblock the barrier.