OFConnectionManager: cleanup unnecessary uses of connection ID
Every time we do a lookup from connection ID to connection_t *, we need to do a
NULL check. But whenever we do have a connection_t *, we know that it is valid
and active. These changes make the code more foolproof by using the pointer
when we already have it.