OFStateManager: add APIs to acquire/release gentable entries
These functions are analogous to the ones defined for the group table. As long
as a reference is held to a gentable entry it can't be deleted, so the private
data pointer remains valid. This allows the referencing code to store the
private data pointer in its own datastructures without risking a dangling
pointer error.