OFConnectionManager: basic bundle implementation
This feature allows the controller to queue up messages to be processed by the
switch. This is useful for two reasons: first, the switch won't be left in an
intermediate state if the controller disconnects while pushing flows; and
second, the controller can send commit messages to every switch at close to the
same instant, minimizing the amount of time the network is in an inconsistent
state.

This doesn't yet implement the full semantics of OF 1.4 bundles. In particular,
a commit is not rolled back if executing a message fails.

Still left to do:
 - Commit in a long-running task
 - Limit maximum bundle size
 - Use a more efficient datastructure than biglist
 - Validate messages on add