OFConnectionManager: add asserts to silence static analysis warning
The cxn_id_to_connection function can usually return null, which would lead to
a null dereference in ind_cxn_accepts_async_message. However, we maintain the
aux_id_to_cxn_id map so that none of the connection IDs are ever stale, so
*cxn will never be NULL.

I don't think this needs to be ported to OFConnectionManager2 since it stores
pointers in aux_id_to_cxn.